requirements.txt
fastapi
uvicorn
streamlit
gradio
langchain
langchain-community
langchain-openai
sentence-transformers
PyMuPDF
jieba
requests
python-dotenv
python-multipart